% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_pdb.R
\docType{bio3d}
\name{bound_pdb}
\alias{bound_pdb}
\title{Bio3D pdb file for bound structure.}
\format{
A bio3d-pdb class object.
\describe{
 \item(xyz){xyz Coordinates of the bound structure for dPPR10-C2 for molecular dynamics.}
}
}
\usage{
data(bound_pdb)
}
\description{
Bio3D pdb file for bound structure.
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{iplotCurves(phe, times)}
}
\references{
Brady Johnston, unpublished work (as of Aug 2020)
}
\keyword{bio3d,}
\keyword{bound,}
\keyword{datasets,}
\keyword{pdb}
