% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_frames.R
\name{compare_frames}
\alias{compare_frames}
\title{Compare Frames}
\usage{
compare_frames(
  bound_pdb,
  apo_pdb,
  frames_model,
  selection,
  string = "calpha",
  ncore = 6,
  fit = T
)
}
\arguments{
\item{bound_pdb}{The bound_pdb reference structure.}

\item{apo_pdb}{The apo_pdb reference structure.}

\item{frames_model}{The bio3d model that contains the frames to be analysed.}

\item{selection}{Numeric AA selection of the model}

\item{string}{Selection string for atom.select.pdb (default is "calpha")}

\item{ncore}{Number of cores to parallelise the workload (defaults to 6). Requires package parallel.}

\item{fit}{Whether or not to fit the structures prior to calculating the RMSD values.}
}
\description{
For comparison within bio3d frame objects. Comparing particular regions
within a structure across a set of frames. Compares to an apo and bound
structures that produces a dataframe with calculated rmsd values.
}
\details{
Returns a dataframe with corresponding frames and rmsds.
}
\examples{

dis_Res(some_pdb, 91, 371)
}
\keyword{FRET,}
\keyword{bio3d,}
\keyword{distance}
\keyword{pdb,}
