% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_res.R
\name{dis_res}
\alias{dis_res}
\title{Distance Between Residues}
\usage{
dis_res(pdb, r1, r2)
}
\arguments{
\item{pdb}{The pdb to be analysed.}

\item{r1}{The first residue.}

\item{r2}{The second residue.}
}
\description{
Uses functions from the Bio3d package to calculate distance of the calphas from two given residues, across all of the frames in a .pdb object.
Returns a dataframe with corresponding frames and distances.
}
\examples{

dis_Res(some_pdb, 91, 371)
}
\keyword{FRET,}
\keyword{bio3d,}
\keyword{distance}
\keyword{pdb,}
