% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_residues.R
\name{distance_residues}
\alias{distance_residues}
\title{Distance Between Residues}
\usage{
distance_residues(some_pdb, res1, res2)
}
\arguments{
\item{some_pdb}{The pdb to be analysed.}

\item{res1}{The first residue.}

\item{res2}{The second residue.}
}
\description{
Uses functions from the Bio3d package to calculate distance of the calphas from two given residues, across all of the frames in a .pdb object.
}
\examples{

distance_residues(some_pdb, 96, 397)
}
\keyword{FRET,}
\keyword{bio3d,}
\keyword{distance}
\keyword{pdb,}
