% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret.R
\name{fret}
\alias{fret}
\title{FRET Calculation}
\usage{
fret(distance, R0 = 54)
}
\arguments{
\item{distance}{Distance of the two fluorophores, in angstroms.}

\item{R0}{R0 radius for the two flouorophores, determins the falloff of the energy transfer.}
}
\description{
Calculates FRET efficiency from input atomic distance and R0 radius.
}
\details{
Using the FRET equaiton, calculates the estimated FRET efficiency from a given distance of two flouorophores. Default R0 value is 54.
}
\examples{

fret(distance = 70)

}
\keyword{fret}
