% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretcalc.R
\name{fretcalc}
\alias{fretcalc}
\title{Calculate FRET}
\usage{
fretcalc(df, R0)
}
\arguments{
\item{df}{The df containing a distance column.}

\item{R0}{The R0 value to be used in the FRET calculation.}
}
\description{
Uses the FRET equation with a given R0 to return expected FRET values for a particular distance column.
}
\examples{
df1 <- data.frame(
  something = c(1,2,3,4),
  distance = rnorm(4, 4, 0.2)
  )

fretcalc(df1, 52)
}
\keyword{FRET}
\keyword{bio3d,}
\keyword{pdb,}
