% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frames.R
\name{get_frames}
\alias{get_frames}
\title{Get Frames From Slices}
\usage{
get_frames(framenos, model, dir = getwd())
}
\arguments{
\item{framenos}{Vector of freames to be selected from the revant slices saved on the HDD.}

\item{model}{Relevant simulaiton to extract frames from.}

\item{dir}{Directory where the relevant frames are saved.}
}
\description{
Returns a pdb object (bio3d) with models from all of the selected frames.
}
\examples{
get_frames(c(123, 45678, 37123), model = "norna")
}
\keyword{bio3d,}
\keyword{extract}
\keyword{frames,}
\keyword{pdb,}
