% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slices.R
\name{get_slices}
\alias{get_slices}
\title{Get Slices}
\usage{
get_slices(frames, divider = 10000)
}
\arguments{
\item{frames}{Vector of the frames to be extracted.}

\item{divider}{Integer for the size of the the slices, in number of frames. Default to 10000.}
}
\description{
Returns a matrix of logicals, for use in filtering in the get_frames function.
}
\examples{
Internal use for the get_frames function.
}
\keyword{bio3d,}
\keyword{filter}
\keyword{pdb,}
