% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_highlight.R
\name{latex_highlight}
\alias{latex_highlight}
\title{Highlight LaTeX}
\usage{
latex_highlight(t1, text, t3, colour = "red")
}
\arguments{
\item{t1}{Text that will precede coloured text. Still monospaced.}

\item{text}{Text that will be highlighted with colour. Still monospaced.}

\item{t3}{Text that will follow coloured text. Still monospaced.}

\item{colour}{Colour of the highlighting. Default "red".}
}
\description{
Returns text with colour and monospaced LaTeX formatting..
}
\details{
Returns text that is surrounded with the \ textcolor and \ texttt wrappers for LaTeX formatting.
}
\examples{

html_highlight("Code\\\\-looking text that ", "will contain red.", "")

}
\keyword{highlight,}
\keyword{html,}
\keyword{monospace}
