% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_highlight.R
\name{mono_highlight}
\alias{mono_highlight}
\title{Mono Highlight HTML}
\usage{
mono_highlight(t1, text, t3, colour = "#FF4500")
}
\arguments{
\item{t1}{Text that will precede coloured text. Still monospaced.}

\item{text}{Text that will be highlighted with colour, and monospaced.}

\item{t3}{Text that will follow coloured text. Still monospaced.}

\item{colour}{Colour of the highlighting. Hex code required i.e. #FF4500}
}
\description{
Returns text with monospaced and colour html formatting.
}
\details{
Returns text that is surrounded with the <tt> html flag for monospaced font, but also additional <strong> and <span style='colour:'#colour'> that will colour the text to emphasise it. Requires special characters like dashes "-" to be be double-escaped "\\-" to be parsed correctly.
}
\examples{

mono_highlight("Code\\\\-looking text that ", "will contain red.", "")

}
\keyword{highlight,}
\keyword{html,}
\keyword{monospace}
