% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_colvar.R
\name{process_colvar}
\alias{process_colvar}
\title{Open and Read ABMD_COLVAR}
\usage{
process_colvar(file, some_colnames = NULL, AAnum = 610)
}
\arguments{
\item{file}{The COLVAR_ABMD file to be opened.}

\item{some_colnames}{Specify other colnames for the read in table other than the default.}

\item{AAnum}{Specify differnt number of amino acid residues for normalising the alpha value, defaulting to 610.}
}
\description{
This function speeds up the openign and processing of the ABMD_COLVAR output file from ABMD inside GROMACS.
}
\examples{
prcoess_colvar(file = "1/COLVAR_ABMD")
}
\keyword{ABMD,}
\keyword{COLVAR,}
\keyword{GROMACS}
