% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomplots.R
\name{random_plots}
\alias{random_plots}
\title{Random Plots}
\usage{
random_plots(x)
}
\arguments{
\item{x}{Number of plots to generate.}
}
\description{
Randomly generates ggplots using {shinyipsum} and adds them together
using \code{patchwork::wrap_plots()}.
}
