% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_comparison.R
\name{region_comparison}
\alias{region_comparison}
\title{Region Comparison}
\usage{
region_comparison(
  bound_pdb,
  apo_pdb,
  frames_model,
  num_segments,
  ncore = 6,
  fit = T
)
}
\arguments{
\item{bound_pdb}{The bound_pdb reference structure.}

\item{apo_pdb}{The apo_pdb reference structure.}

\item{frames_model}{The bio3d model that contains the frames to be analysed.}

\item{num_segments}{Number of segments to split the AA chain into.}

\item{ncore}{Number of cores to parallelise the workload (defaults to 6). Requires package parallel.}

\item{fit}{Whether or not to fit the structures prior to calculating the RMSD values.}
}
\description{
Compares a specfic section of a structure (region) along the entirety of the
structure across multiple frames in a bio3d object.
}
\details{
Returns a dataframe with corresponding frames, rmsds and regions.
}
\examples{

}
\keyword{FRET,}
\keyword{bio3d,}
\keyword{distance}
\keyword{pdb,}
