% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_xyz.R
\name{res_xyz}
\alias{res_xyz}
\title{Residue XYZ}
\usage{
res_xyz(pdb, res, atomtype = "calpha")
}
\arguments{
\item{pdb}{The pdb to be processed.}

\item{res}{The dbl index of the selected residue.}

\item{atomtype}{Selection of atom to be return, defaults to calpha.}
}
\description{
Returns the XYZ coordinate of the specified atom from the selected residue, across all frames of a pdb object.
}
\examples{
res_xyz(some_pdb, 6)
}
\keyword{bio3d,}
\keyword{calpha}
\keyword{pdb,}
