% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_table.R
\docType{data}
\name{rna_table}
\alias{rna_table}
\title{Data Table of Experimental RNA}
\format{
A data frame with columns:
\describe{
 \item(name){Name given to the sequnce, BAJ_01-BAJ_13 and noRNA.}
 \item(latex_name){Name that has escaped underscores "\_" for LaTeX formatting.}
 \item(raw_seq){RNA sequence, 5ʹ to 3ʹ of each RNA.}
 \item(html_seq){Sequences with mismatch to BAJ_01 highilgihted in RED using html flags. Compatible with geom_richtext for labelling.}
 \item(mono_seq){Seqeunces with mismatches to BAJ_01 highlighed in RED and also a monospaced font, using html flags. Compatible with kable(html) for making tables.}
 \item(latex_seq){Sequences with misatch to BAJ_01 highlighed in RED with LaTeX formatting and also a monospaced font. Compatible with kable(latex) for makint tables.}
 }
}
\usage{
data(rna_table)
}
\description{
Sequences of RNA for MD and FRET experiments. Raw sequences are available, as well as some
pre-formatted sequences for use in geom_richtext and kable for both html and LaTeX, that highlight
the mismatches between these and the target RNA sequence.
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{iplotCurves(phe, times)}
}
\references{
Brady Johnston, unpublished work (as of Aug 2020)
}
\keyword{datasets,}
\keyword{rna,}
\keyword{sequnce}
