% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_comparison.R
\name{sliding_comparison}
\alias{sliding_comparison}
\title{Sliding Comparison}
\usage{
sliding_comparison(
  segment_size,
  segment_slide,
  bound_pdb,
  apo_pdb,
  frames_model,
  ncore = 6,
  fit = T
)
}
\arguments{
\item{segment_size}{Size of AA sequence in the comparison segment.}

\item{segment_slide}{Change in segment comparison, per comparison.}

\item{bound_pdb}{The bound_pdb reference structure.}

\item{apo_pdb}{The apo_pdb reference structure.}

\item{frames_model}{The bio3d model that contains the frames to be analysed.}

\item{ncore}{Number of cores to parallelise the workload (defaults to 6). Requires package parallel.}

\item{fit}{Whether or not to fit the structures prior to calculating the RMSD values.}
}
\description{
Takes an AA region of size segment_size, and compares between structures
entirety of AA sequence. Moves segment by segment_slide each time.
}
\details{
Returns a dataframe with corresponding frames, rmsds and regions.
}
\examples{

}
\keyword{FRET,}
\keyword{bio3d,}
\keyword{distance}
\keyword{pdb,}
