% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\docType{csv}
\name{spectrum}
\alias{spectrum}
\title{Data frame of visible light values.}
\format{
Data frame with rgb data for visible spectrum.
}
\usage{
data(spectrum)
}
\description{
Data frame of visible light values.
}
\references{
Brady Johnston, unpublished work (as of Aug 2020)
}
\keyword{datasets,}
\keyword{nm}
\keyword{visible,}
\keyword{wavelength,}
